backend_app/requirements.txt
# Django
asgiref==3.2.10
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
Django==3.1.1
django-admin==2.0.1
django-common-helpers==0.9.2
django-cors-headers==3.10.1
django-cron==0.5.1
django-debug-toolbar==3.2.2
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.5
django-filter==21.1
django-rq==2.5.1
django-six==1.0.4
djangorestframework==3.12.4
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
psycopg2==2.9.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
rq==1.10.1
screen==1.0.1
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.3
xlwt==1.3.0
zipp==3.6.0

# Flask
Flask==1.1.2
SQLAlchemy==1.3.20
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Cors==3.0.9
mysqlclient==2.0.1
pika==1.1.0

# ML
cycler==0.11.0
fonttools==4.28.3
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing_extensions==4.0.1
